requirements.txt
streamlit>=1.28.0
primer3-py>=2.0.0
biopython>=1.80
pandas>=1.5.0
plotly>=5.0.0
requests>=2.28.0